change enable_fips to gracefully complete on success. Other changes to update scripts

+++ security/crypto/enable_fips/runtest.sh
@@ -31,21 +31,24 @@ rlJournalStart
 
     rlPhaseStartTest
         rlRun "fips-mode-setup --check"
-        rlRun "fips-mode-setup --is-enabled" && rlPass "FIPS mode is enabled" && exit 0
-        [[ -e /tmp/enable_fips_attempted ]] && rlDie "Failed to enable FIPS"
-        touch /tmp/enable_fips_attempted && sync
-        if stat /run/ostree-booted > /dev/null 2>&1; then
-            rlRun "fips-mode-setup --enable --no-bootcfg"
-            kernel_args=$(fips-mode-setup --enable --no-bootcfg | awk -F\" '/fips=1/ {print $2}')
-            kernel_current=$(grubby --info=DEFAULT | awk -F\" '/kernel=/ {print $2}')
-            grubby --update-kernel="${kernel_current}" --args="${kernel_args}"
-        else
-            rlRun "fips-mode-setup --enable"
+        rlRun "fips-mode-setup --is-enabled" && rlPass "FIPS mode is enabled" && fips_enabled=1
+        if [ ! ${fips_enabled} ]; then
+            [[ -e /tmp/enable_fips_attempted ]] && rlDie "Failed to enable FIPS"
+            touch /tmp/enable_fips_attempted && sync
+            if stat /run/ostree-booted > /dev/null 2>&1; then
+                rlRun "fips-mode-setup --enable --no-bootcfg"
+                kernel_args=$(fips-mode-setup --enable --no-bootcfg | awk -F\" '/fips=1/ {print $2}')
+                kernel_current=$(grubby --info=DEFAULT | awk -F\" '/kernel=/ {print $2}')
+                grubby --update-kernel="${kernel_current}" --args="${kernel_args}"
+            else
+                rlRun "fips-mode-setup --enable"
+            fi
+            rlRun "rhts-reboot"
         fi
-        rlRun "rhts-reboot"
     rlPhaseEnd
 
     rlPhaseStartCleanup
+    [[ -e /tmp/enable_fips_attempted ]] && rlRun "rm /tmp/enable_fips_attempted"
     rlPhaseEnd
 rlJournalEnd
 rlJournalPrintText
